CHOR - add interface for userspace program (VADD, VMUL, GEMV) via IOCTLs

+++ include/pim_memory_region.h
@@ -17,6 +17,6 @@
 
 
 extern volatile u32 __iomem *pim_data_virt_addr;
-extern volatile u32 __iomem *pim_config_virt_addr;
+extern volatile u8 __iomem *pim_config_virt_addr;
 
 #endif
